pyproject.toml
[tool.poetry]
name = "exam_fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.39"
fastapi = "^0.78.0"
uvicorn = "^0.18.3"
asyncpg = "^0.26.0"
PyYAML = "^6.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pyproject-flake8 = "^0.0.1-alpha.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ['py310']
include = '\.py$'

[tool.isort]
profile = "black"
line_length = 119

[tool.flake8]
max-line-length = 119
max-complexity = 18
ignore = "E203,E266,W503,"

[tool.pytest.ini_options]
testpaths = ["tests",]
filterwarnings = ["ignore::DeprecationWarning",]